Añadida nueva cuenta objetivos y modelo objetivos

+++ app/Http/Controllers/AccountController.php
@@ -7,45 +7,59 @@ use App\Http\Requests\StoreAccountRequest;
 use App\Http\Requests\UpdateAccountRequest;
 use App\Interfaces\AccountRepositoryInterface;
 use App\Models\Account;
+use Illuminate\Contracts\View\View;
+use Illuminate\Http\RedirectResponse;
 
 class AccountController extends Controller
 {
-    private AccountRepositoryInterface $accountRepository;
-
-    public function __construct(AccountRepositoryInterface $accountRepository)
-    {
-        $this->accountRepository = $accountRepository;
-    }
+    public function __construct(
+        private AccountRepositoryInterface $accountRepository
+    ) {}
 
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return View
      */
-    public function index()
+    public function index(): View
     {
         return view('accounts.index', [
-            'accounts'   => $this->accountRepository->getAllAccounts(),
-            'titleRight' => 'Cuentas'
+            'accounts'     => $this->accountRepository->getAccountsByType('Normal'),
+            'goalAccounts' => $this->accountRepository->getAccountsByType('Objetivos'),
+            'titleRight'   => 'Cuentas'
+        ]);
+    }
+
+    /**
+     * Display the specified resource.
+     *
+     * @param Account $account
+     *
+     * @return View
+     */
+    public function show(Account $account): View
+    {
+        $this->authorize('view', $account);
+
+        if ($account->type == 'Objetivos') {
+            return view('accounts.goals.show', [
+                'account'    => $account,
+                'titleRight' => $account->name,
+            ]);
+        }
+
+        return view('accounts.show', [
+            'account'    => $account,
+            'titleRight' => $account->name,
         ]);
     }
 
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return View
      */
-    public function show($accountId)
-    {
-        $account = $this->accountRepository->getAccountById($accountId);
-
-        return view('accounts.show', [
-            'account'       => $account,
-            'titleRight'    => $account->name,
-        ]);
-    }
-
-    public function create()
+    public function create(): View
     {
         return view('accounts.form', [
             'method'     => 'POST',
@@ -54,7 +68,14 @@ class AccountController extends Controller
         ]);
     }
 
-    public function store(StoreAccountRequest $request)
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @param StoreAccountRequest $request
+     *
+     * @return RedirectResponse
+     */
+    public function store(StoreAccountRequest $request): RedirectResponse
     {
         $account = $this->accountRepository->createAccount(
             $request->only([
@@ -63,6 +84,7 @@ class AccountController extends Controller
                 'balance',
                 'category_id',
                 'main',
+                'type',
             ])
         );
 
@@ -70,19 +92,28 @@ class AccountController extends Controller
             $account,
             'account',
             'Cuenta creada',
-            'Se ha creado la cuenta ' . $account->name,
+            'Se ha creado la cuenta '.$account->name,
             route('accounts.show', $account->id)
         ));
 
-        return redirect()->route('accounts.index');
+        return redirect()->route('accounts.show', $account->id);
     }
 
-    public function edit($accountId)
+    /**
+     * Show the form for editing the specified resource.
+     *
+     * @param Account $account
+     *
+     * @return View
+     */
+    public function edit(Account $account): View
     {
+        $this->authorize('update', $account);
+
         return view('accounts.form', [
-            'account'    => $this->accountRepository->getAccountById($accountId),
+            'account'    => $account,
             'method'     => 'PUT',
-            'route'      => route('accounts.update', $accountId),
+            'route'      => route('accounts.update', $account->id),
             'titleRight' => 'Editar cuenta'
         ]);
     }
@@ -90,12 +121,15 @@ class AccountController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Account  $account
-     * @return \Illuminate\Http\Response
+     * @param  UpdateAccountRequest $request
+     * @param  Account              $account
+     *
+     * @return RedirectResponse
      */
-    public function update(UpdateAccountRequest $request, Account $account)
+    public function update(UpdateAccountRequest $request, Account $account): RedirectResponse
     {
+        $this->authorize('update', $account);
+
         $this->accountRepository->updateAccount(
             $account,
             $request->only([
@@ -118,8 +152,17 @@ class AccountController extends Controller
         return redirect()->route('accounts.index');
     }
 
-    public function destroy(Account $account)
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param Account $account
+     *
+     * @return RedirectResponse
+     */
+    public function destroy(Account $account): RedirectResponse
     {
+        $this->authorize('delete', $account);
+
         $this->accountRepository->deleteAccount($account);
 
         event(new ActivityEvent(
